moved reg_peaks to segfunctions
